# Conflicts:
#	modules/gdscript/gd_function.cpp
#	modules/gdscript/gd_parser.cpp
#	modules/gdscript/gd_script.cpp
#	modules/gdscript/register_types.cpp

+++ servers/audio/audio_filter_sw.cpp
@@ -5,7 +5,7 @@
 /*                           GODOT ENGINE                                */
 /*                    http://www.godotengine.org                         */
 /*************************************************************************/
-/* Copyright (c) 2007-2016 Juan Linietsky, Ariel Manzur.                 */
+/* Copyright (c) 2007-2017 Juan Linietsky, Ariel Manzur.                 */
 /*                                                                       */
 /* Permission is hereby granted, free of charge, to any person obtaining */
 /* a copy of this software and associated documentation files (the       */
@@ -142,9 +142,9 @@ void AudioFilterSW::prepare_coefficients(Coeffs *p_coeffs) {
 			//this one is extra tricky
 			double hicutoff=resonance;
 			double centercutoff = (cutoff+resonance)/2.0;
-			double bandwidth=(Math::log(centercutoff)-Math::log(hicutoff))/Math::log(2);
+			double bandwidth=(Math::log(centercutoff)-Math::log(hicutoff))/Math::log((double)2);
 			omega=2.0*Math_PI*centercutoff/sampling_rate;
-			alpha = Math::sin(omega)*Math::sinh( Math::log(2)/2 * bandwidth * omega/Math::sin(omega) );
+			alpha = Math::sin(omega)*Math::sinh( Math::log((double)2)/2 * bandwidth * omega/Math::sin(omega) );
 			a0=1+alpha;
 
 			p_coeffs->b0 =  alpha;
